+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Bug fixes
 - Fixed stopping services in `started` state.
 
+### Improvements
+- Send request data to sentry, add tag ship_IP to events.
+
 ## 1.8.1 (2016-11-03)
 
 We do best effort to support docker versions 1.6.0 - 1.12.1 with this release.